move code from init to new main file to prevent double initialization of globals
